IMF and Wd1 changes on laptop.

+++ jlu/imf/imf.py
@@ -61,10 +61,10 @@ def sample_imf(massLimits, imfSlopes, totalMass,
     if not makeMultiples:
         newStarCount *= 1.1
 
-    masses = np.array([], dtype=float)
+    masses = np.array([], dtype=np.float32)
     isMultiple = np.array([], dtype=bool)
     compMasses = []
-    systemMasses = np.array([], dtype=float)
+    systemMasses = np.array([], dtype=np.float32)
 
     loopCnt = 0
 
@@ -166,7 +166,7 @@ class IMF_broken_powerlaw(object):
 
         # Calculate the coeffs to make the function continuous
         nterms = len(self.powers)
-        coeffs = np.ones(nterms, dtype=float)
+        coeffs = np.ones(nterms, dtype=np.float32)
 
         # First term is just 1.0
         # Subsequent terms are products of previous terms and then some.
@@ -194,7 +194,7 @@ class IMF_broken_powerlaw(object):
         returnFloat = type(m) == float
         
         m = np.atleast_1d(m)
-        xi = np.zeros(len(m), dtype=float)
+        xi = np.zeros(len(m), dtype=np.float32)
         
         for i in range(len(xi)):
             tmp = gamma_closed(m[i], self.mLimitsLow, self.mLimitsHigh)
@@ -214,7 +214,7 @@ class IMF_broken_powerlaw(object):
         """
         returnFloat = type(m) == float
         m = np.atleast_1d(m)
-        mxi = np.zeros(len(m), dtype=float)
+        mxi = np.zeros(len(m), dtype=np.float32)
         
         for i in range(len(mxi)):
             tmp = gamma_closed(m[i], self.mLimitsLow, self.mLimitsHigh)
@@ -245,7 +245,7 @@ class IMF_broken_powerlaw(object):
         returnFloat = type(a) == float
 
         a = np.atleast_1d(a)
-        val = np.zeros(len(a), dtype=float)
+        val = np.zeros(len(a), dtype=np.float32)
 
         for i in range(len(val)):
             t1 = theta_open(a[i] - self.mLimitsHigh) * self.coeffs
@@ -270,7 +270,7 @@ class IMF_broken_powerlaw(object):
         returnFloat = type(a) == float
         
         a = np.atleast_1d(a)
-        val = np.zeros(len(a), dtype=float)
+        val = np.zeros(len(a), dtype=np.float32)
 
         for i in range(len(val)):
             t1 = theta_open(a[i] - self.mLimitsHigh) * self.coeffs
@@ -392,8 +392,8 @@ class IMF_broken_powerlaw(object):
         r = np.atleast_1d(r)  # Make sure it is an array
         
         x = r * self.lamda[-1]
-        y = np.zeros(len(r), dtype=float)
-        z = np.ones(len(r), dtype=float)
+        y = np.zeros(len(r), dtype=np.float32)
+        z = np.ones(len(r), dtype=np.float32)
 
         for i in range(self.nterms):
             aux = x - self.lamda[i]
@@ -445,7 +445,7 @@ class IMF_Chabrier_2003(object):
 
         # Calculate the coeffs to make the function continuous
         nterms = len(self.powers)
-        coeffs = np.ones(nterms, dtype=float)
+        coeffs = np.ones(nterms, dtype=np.float32)
 
         # First term is just 1.0
         # Subsequent terms are products of previous terms and then some.
@@ -473,7 +473,7 @@ class IMF_Chabrier_2003(object):
         returnFloat = type(m) == float
         
         m = np.atleast_1d(m)
-        xi = np.zeros(len(m), dtype=float)
+        xi = np.zeros(len(m), dtype=np.float32)
         
         for i in range(len(xi)):
             tmp = gamma_closed(m[i], self.mLimitsLow, self.mLimitsHigh)
@@ -493,7 +493,7 @@ class IMF_Chabrier_2003(object):
         """
         returnFloat = type(m) == float
         m = np.atleast_1d(m)
-        mxi = np.zeros(len(m), dtype=float)
+        mxi = np.zeros(len(m), dtype=np.float32)
         
         for i in range(len(mxi)):
             tmp = gamma_closed(m[i], self.mLimitsLow, self.mLimitsHigh)
@@ -524,7 +524,7 @@ class IMF_Chabrier_2003(object):
         returnFloat = type(a) == float
 
         a = np.atleast_1d(a)
-        val = np.zeros(len(a), dtype=float)
+        val = np.zeros(len(a), dtype=np.float32)
 
         for i in range(len(val)):
             t1 = theta_open(a[i] - self.mLimitsHigh) * self.coeffs
@@ -549,7 +549,7 @@ class IMF_Chabrier_2003(object):
         returnFloat = type(a) == float
         
         a = np.atleast_1d(a)
-        val = np.zeros(len(a), dtype=float)
+        val = np.zeros(len(a), dtype=np.float32)
 
         for i in range(len(val)):
             t1 = theta_open(a[i] - self.mLimitsHigh) * self.coeffs
@@ -671,8 +671,8 @@ class IMF_Chabrier_2003(object):
         r = np.atleast_1d(r)  # Make sure it is an array
         
         x = r * self.lamda[-1]
-        y = np.zeros(len(r), dtype=float)
-        z = np.ones(len(r), dtype=float)
+        y = np.zeros(len(r), dtype=np.float32)
+        z = np.ones(len(r), dtype=np.float32)
 
         for i in range(self.nterms):
             aux = x - self.lamda[i]
@@ -910,7 +910,7 @@ def theta_closed(x):
     isFloat = type(x) == float
 
     x = np.atleast_1d(x)
-    val = np.ones(len(x), dtype=float)
+    val = np.ones(len(x), dtype=np.float32)
     val[x < 0] = 0.0
 
     if isFloat:
@@ -926,7 +926,7 @@ def theta_open(x):
     isFloat = type(x) == float
 
     x = np.atleast_1d(x)
-    val = np.zeros(len(x), dtype=float)
+    val = np.zeros(len(x), dtype=np.float32)
     val[x > 0] = 1.0
 
     if isFloat:
@@ -942,7 +942,7 @@ def delta(x):
     isFloat = type(x) == float
 
     x = np.atleast_1d(x)
-    val = np.ones(len(x), dtype=float)
+    val = np.ones(len(x), dtype=np.float32)
     val[x == 0] = 0.5
 
     if isFloat:
